requirements.txt
google-api-python-client
google-auth
flask
openai